In the middle of a refactoring of the go modules. This is not a viable solution, but submitting to branch to share with collaborators.
