--- github.com/go.mod
@@ -1,3 +1,0 @@
-module github.com/ostelco-core
-
-go 1.12
