+++ sim-administration/sim-batch-management/loltelutils/loltelutils.go
@@ -1,0 +1,24 @@
+package loltelutils
+
+
+func TrimSuffix(s string, suffixLen int) string {
+	return s[:len(s)-suffixLen]
+}
+
+func Sign(x int) int {
+	if x < 0 {
+		return -1
+	} else if x > 0 {
+		return 1
+	} else {
+		return 0
+	}
+}
+
+// Abs returns the absolute value of x.
+func Abs(x int) int {
+	if x < 0 {
+		return -x
+	}
+	return x
+}
