+++ sim-administration/sim-batch-management/outfileconversion/outfile_to_hss_input_converter_lib_test.go
@@ -1,4 +1,4 @@
-package main
+package outfileconversion
 
 import (
 	"gotest.tools/assert"
@@ -14,7 +14,7 @@ func testKeywordValueParser(t *testing.T) {
 
 func testReadOutputFile(t *testing.T) {
 	sample_output_file_name := "sample_out_file_for_testing.out"
-	record, _ := ReadOutputFile(sample_output_file_name)
+	record := ReadOutputFile(sample_output_file_name)
 
 	// First parameter to check
 	assert.Equal(t, sample_output_file_name, record.Filename)
