+++ sim-administration/sim-batch-management/uploadtoprime/luhn_test.go
@@ -1,4 +1,4 @@
-package main
+package uploadtoprime
 
 
 import (
